requirements.txt
crispy-bootstrap5==2023.10
Django==4.2.7
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-import-export==3.3.3
Pillow==10.1.0
python-slugify==8.0.1
slugify==0.0.1
